+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+## [1.1.0] (Agust 8, 2019)
+- Named interpolation
+- added an example
+
 ## [1.0.1] (Agust 6, 2019)
-- nullsafety for locales
+- Null safety for locales
 
 ## [1.0.0+1] (July 27, 2019)
 - Refactor code